Experiment with moving local support libraries into a single lib file.
This could be ever so slightly faster than distributing in multiple parts.

It also adds a small degree of privacy in that sublibs cannot be loaded on their own. ie, they can only be imported after the main lib has been loaded.

There may be issues in compiling programs using this layout, hence the commit to a branch.