Remove guardian stuff, ev-free-watcher! and TODO section from README.
The guardian experiment turned out not to be practical with how ev programs are structured.

ie, they always keep references to watchers or ignore them (in which case we don't want guardians to auto clean them).

And undefining a watcher reference is less clear than calling ev-TYPE-free explicitly.

ev-free-watcher! was mostly useful for guardian cleanup, so that is removed in favour of clients using type specific free functions.

The TODO section is removed entirely as that contained mention of guardian support and a number of other stale ideas.